[sw/sca] Add simple serial library
This change introduces a library that implements simple serial protocol
version 1.1. This library provides:
- A mechanism for registering command handlers,
- A simple parser that parses incoming packets and dispatches them to
  the appropriate handlers,
- A built-in handler for the 'v' (version) command, and
- Functions to send arbitrary packets to the host.
